Support for name querying.
Example: `(Query (() None PpStr) (Names "nat_"))`

I'm far from happy with this, as it will ignore longer aliases for
constants, but it uses a similar method than the one in company Coq and
should be way faster; so people can play a bit with completion if they want.

The general question remains open: it seems the `Nametab` is the only
part with full capability for providing good completion, but it remains sealed.