[sertop] [async] Fix connection to workers.
Coq's worker manager attemps to connect to workers using our command
line; however this fails as out command line is not compatible with coqtop.

Thus, we fix our command line for now and extend the async parameter to
take the location of the desired worker toplevel.

With this setup, you can do:
```
./sertop.native --prelude ~/external/coq-git/ --async ~/external/coq-git/bin/coqtop
```
and it seems to process files such as `test-suite/interactive/proof-block.v`
fully asynchronously. This makes significant progress on issue #13.