[proto] Towards a serialization-independent core.
Even if at some boilerplate code, it makes a good discipline exercise to
maintain the core free of a particular serialization choice.

We do however document some choices in the header as `ppx_conv_sexp` is
still our main target.

This also allows for a better integration with JSON-native parts like
the `jslib` manager.

A more refined option would be to parameterize the `CoqObject` type by
the serialization type; but IMO this is overkill.