[sertop] Ensure initial state is 1, even when loading the prelude.
Command line options shouldn't affect the initial state, we guarantee it
to be 1 after STM initialization. Thus, we make the prelude option to
atomically load the prelude before the STM init.

If the user wants to have more control over it, it can just add the
desired `Require Import` manually.

Refs #14, but we need to close:

- Import the `Cancel 1` handling from jsCoq.
- Provide a `--no-init` option plus an Init control command.