[serapi] Extend CoqExn information
_warning_: Protocol-breaking change.

The error handling strategy for STM clients seems not to be fully
specified. In particular, it seems that clients are required to handle
synchronous exceptions in addition to asynchronous error feedback, as
such is not always generated for some cases.

CoqIDE/XML protocol indeed fully ignores "error feedback" and provides a
custom, synchronous `Fail` answer forwarding key exception information.
In particular, `Fail` includes some state information that allows to
track the precise location of the exception and recover appropriately.

Thus, we extend `CoqExn` to incorporate state and location information
such that SerAPI clients can handle all the synchronous errors properly.