[serlib] Handle serialization of generic arguments. [see #41]
In previous versions of SerAPI, we used to handle generic arguments in
an ad-hoc way, but that was a pain to implement.

We now handle serialization of generic arguments using Coq's register
mechanism. Thus, parts of the system can register serializers for
generic arguments.

We have registered quite a few handlers, but more remain.

This should solve the first part of #41, however, de-serialization is
still not implemented, thus keeping the bug open.