[sertop] Support command line options `-R` and `-Q`.
Now users can do:
```
sertop -Q lp,dir
```

to bind a logical path `lp` to directory `dir` in the same way as Coq does.

Note that support is still incomplete and experimental, in particular
we don't properly support recursive scanning of `dir`. Also, using an
empty space in the option is not possible due to `cmdliner`
limitations. We have chosen a comma as separator and indeed this looks
fine to me.

c.f. #19, #33, #35.