[sertop] Refactor load path handling.
We add a more flexible interface for load path initialization. This
will allow us to support for the traditional model [based on recursive
load path with recursive directory scanning] and the "web" model,
where a unix filesystem is not available and we must provide the full,
non-recursive loadpath.

This interface is a candidate to be ported upstream.