[serlib] Some more progress towards generic argument de-serialization.
c.f. #41. **Warning** this changes the serialization format of generic
arguments, which now looks as:

```lisp
(GenArg raw (OptArg (ExtraArg ltac_selector)) ())
```

In general, the format is: `(GenArg $lvl $type $obj)` where `$lvl` is
the "level" of the generic argument, `$type` is the generic type AST
and `$obj` the serialized object proper.

This format should remain quite stable in the near future.