[serlib] Fully generic argument serialization/deserialization.
This closes #41. We had to do a bit of magic to get deserialization
working, I'm not sure it can be avoided easily.

Also, we have a very clunky and verbose scheme to register generic
arguments. We should improve it, however, as far as I know `deriving'
doesn't still have the capability to automatically generate
serializers for a set of types or modules, so this seems hard to avoid
for now.