[serapi] Add `Assumptions` query.
```
(Query () (Assumptions global)
```

will now return an assumptions object for `global`. Usual query
options do apply.

Support for this query is experimental as the returned object is not
stable.

We have also taken the opportunity to perform a cleanup of handling of
the environment, we now don't use the `Global.*` family of functions.