[serlib] [sertop] Load plugin serializers at plugin loading time.
We load serializers for plugin's generic arguments at plugin loading
time. This avoid having some conflicts as indeed, as the number of
plugins grew large, static linking didn't scale.

This fixes the issue reported in #90

Note that tests must now depend on the full `coq-serapi` package as
for the plugins to be available to `Fl_dynload`.