[serapi] Hook into new 8.10 complete API
This is not very useful at the moment, as it is means more as a
placeholder to test than as a device to perform real completion,
however it is much nicer than the current no-support.

c.f. https://github.com/coq/coq/pull/8766
c.f. https://github.com/coq/coq/issues/7164