[serlib] Fix ppx_import w.r.t. old-style pack plugin compilation
This is necessary when Coq is compiled with the standard Makefile.

In order to have a solution that both works with `-pack` and Dune
mode, we need to wrap the mli-only file that Coq is shipping upstream
due to differences in linking.

Fixes #143

We also update the Travis setup so we test both OPAM and the Dune build.