[serapi] New command `SaveDoc` to save .vo files.
We introduce a new commmand:
```lisp
(SaveDoc opts)
```
to save .vo files from an interactive session. `opts` are optional,
and can be

- `(sid s)`: save the document at a specific sentence
- `(prefix_output_dir dir)` add a prefix to the `file.vo` to be saved.

Note that Coq requires that the module name and the filename do match,
thus in order to use this command `sertop` must be called with the
`--topfile file.v` flag, otherwise `CannotSaveVo` will be raised.

Closes #128 #238