[serapi] Support Coq 8.14
Nothing significant it seems, other than a bit of a mess with
initialization due to upstream changes. There is much room for
improvement if we get time to refactor. We also have to account for
`coq-core` until we improve plugin location using `findlib`.

Also, some plugins were renamed, which helps us for consistency in our
plugin loader.