[serlib] Ignore `env` parameter in certain exceptions
Should fix #250, tho it is far from a principled solution.

Several Coq structures do like to embed a full `env`; what we do here,
it is far from practical, as it won't scale, but at least should
alleviate the immediate pain

Other possible hack would to add an `--omit_env` parameter that would
turn all `env` into abstract; but generally, we should really
introduce a mechanism for the management of contextual info as
discussed in the bug tracker.